refactor(refs T30969): move DpAccordion to components root
