--- src/components/core/DpAccordion.stories.mdx
@@ -1,14 +1,0 @@
-# Accordion
-
-The accordion component can toggle itself via the toggle method, but can also be toggled from outside
-(e.g. via a save or abort button) by passing the parent's open state as prop (isOpen) and calling
-a toggle method on the parent @item:toggle which toggles the parent's isOpen data property.
-
-## With toggle from outside
-
-```
-<dp-accordion
-  :is-open="isOpen"
-  :title="title"
-  @item:toggle="(open) => toggleItem(open)" />
-```
